css-require-hook works with stylable and css-modules
