+++ package.json
@@ -10,7 +10,7 @@
     "prepublish": "npm run build",
     "test": "npm run test:browser && npm run test:node",
     "test:browser": "karma start --single-run",
-    "test:node": "mocha --compilers ts:ts-node/register,tsx:ts-node/register,css:test/stylable-register \"./test/node/**/*.node-spec.ts?(x)\"",
+    "test:node": "mocha --compilers ts:ts-node/register,tsx:ts-node/register,css:test/utils/css-require-hook \"./test/node/**/*.node-spec.ts?(x)\"",
     "start": "webpack-dev-server --progress"
   },
   "peerDependencies": {
