--- test/stylable-register.js
@@ -1,4 +1,0 @@
-const fs = require('fs');
-const hook = require('stylable-integration/dist/src/require-hook')
-
-hook.attachHook({extension: '.css'})
